+++ tests/codegen_testcases/solidity/common_subexpression_elimination.sol
@@ -299,7 +299,7 @@ contract c1 {
     function test11(int a, int b) public returns (int) {
         string ast = "Hello!";
         string bst = "from Solang";
-        string cst = ast + bst;
+        string cst = string.concat(ast, bst);
         // CHECK: ty:int256 %1.cse_temp = (signed divide (arg #0) / (int256 2 * (arg #1)))
         // CHECK: call c1::c1::function::get__int256_int256 %1.cse_temp, (arg #1)
         int p = a + get(a/(2*b), b);
@@ -308,22 +308,22 @@ contract c1 {
         // CHECK: ty:bool %2.cse_temp = (strcmp (%ast) (%bst))
         // CHECK: branchcond %2.cse_temp, block2, block1
         bool e2 = e;
-        // CHECK: branchcond (strcmp (%cst) (%cst)), block3, block4
-        if (ast + bst == cst) {
+        // CHECK: branchcond (strcmp ((builtin Concat (%ast, %bst))) (%cst)), block3, block4
+        if (string.concat(ast, bst) == cst) {
             // CHECK: call c1::c1::function::get__int256_int256 %1.cse_temp, (arg #1)
             require(a + get(a/(2*b), b) < 0);
-            emit testEvent(a + get(a/(2*b) -p, b), p, ast+bst);
+            emit testEvent(a + get(a/(2*b) -p, b), p, string.concat(ast, bst));
         }
 
         // CHECK: branchcond %2.cse_temp, block21, block22
         if (ast == bst) {
-            ast = ast + "b";
+            ast = string.concat(ast, "b");
         }
         // CHECK: call c1::c1::function::get__int256_int256 (%1.cse_temp - %p), (arg #1)
 
         // CHECK: branchcond (strcmp (%ast) (%bst)), block24, block25
         while (ast == bst) {
-            ast = ast + "a";
+            ast = string.concat(ast, "a");
         }
 
         // CHECK: call c1::c1::function::get__int256_int256 (arg #1), (signed divide (arg #0) / (arg #1))
@@ -378,17 +378,17 @@ contract c1 {
         if(vec.length - (a+b) == 1) {
             // CHECK:  call c1::c1::function::testing__bytes %c
             string k = testing(bytes(c));
-            string p = "a" +k;
-            // CHECK: ty:string %p = (concat ((alloc string uint32 1 "a")) (%k))
+            string p = string.concat("a", k);
+            // CHECK: ty:string %p = (builtin Concat ((alloc string uint32 1 "a"), %k))
             // CHECK: branchcond ((builtin ArrayLength (%p)) == uint32 2), block11, block12
             if(p.length == 2) {
-                // CHECK: ty:string %p1 = (concat ((alloc string uint32 1 "a")) (%k))
-                string p1 = "a" + k;
+                // CHECK: ty:string %p1 = (builtin Concat ((alloc string uint32 1 "a"), %k))
+                string p1 = string.concat("a", k);
                 string l = p1;
             }
         }
 
-        // CHECK: branchcond (signed less (%a + (arg #1)) < int256 0), block14, block15
+        // CHECK: branchcond (signed less %2.cse_temp < int256 0), block14, block15
         while(a+b < 0) {
             // CHECK: branchcond (strcmp (%c) ("a")), block16, block17
             if("a" == c) {